backend/api/requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-localflavor==3.0.1
django-localflavor-us==1.1
django-location-field==2.1.0
django-mysql==3.8.1
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-extensions==0.6.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
phonenumbers==8.12.6
psycopg2==2.8.5
PyJWT==1.7.1
python-stdnum==1.13
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
